Fixed allocation size mismatches for saved rendition data.  Fixed several memory issues per valgrind.

+++ repaint.h
@@ -10,6 +10,6 @@ void repaint(int row1, int row2, int col1, int col2);
 
 //  Paint the text using the rendition value at the screen position.
 void paint_rval_text(unsigned char * restrict str, uint32_t rval,
-	unsigned int len, int x, int y);
+	uint8_t len, int16_t x, int16_t y);
 
 #endif//JBXVT_REPAINT_H
