+++ slinest.h
@@ -7,7 +7,7 @@
 struct slinest {
 	unsigned char *sl_text;	/* the text of the line */
 	uint32_t *sl_rend;	/* the rendition style */
-	int sl_length;	/* length of the line */
+	uint8_t sl_length;	/* length of the line */
 };
 
 #endif//!SLINEST_H
